Add files via upload
Briefly summarize the DriverPass project. Who was the client? What type of system did they want you to design?
The DriverPass system was a project that was created for a diving school so new students and clients could be reached to take practice tests, stay up to date on driving rules and regulations from DMV and schedule drivers training. They wanted an easy accessible system from multiple platforms and devices that was cloud based. 
What did you do particularly well?
I think I planned the project well and thought of things that weren't easily thought of. Portions of the coding were easier than others. 
If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it?
I think I would make some of the sequence diagrams a little more easy to follow. I tend to think a lot and over analyze so I think some diagrams I put more information than was necessary which can cause time delay with the customer because of questions and need for clarification. 
How did you interpret the user’s needs and implement them into your system design? Why is it so important to consider the user’s needs when designing?
I made sure I included everything in the system that the customer wanted. There were some restraints so not everything was possible and that was something that was put into the restraints of the system and how much a team could do and accomplish within a timeframe given. 
How do you approach designing software? What techniques or strategies would you use in the future to analyze and design a system?
Ik approach designing software like building a building. Starting with the thoughts and basic needs, then blueprints, then revising blueprints, then final sketches and beginning the work. When you hit a hurdle or snag we go back and revise then proceed until we have a solid foundation then work on aesthetics.